Change the type of SparseMatrix::num_nonzeroes and SparseMatrix::max_num_nonzeroes from int to int64_t.
This will allow for larger size Jacobians to be allocated, when memory
is available and matrix dimensions are still 32bit.

https://github.com/ceres-solver/ceres-solver/issues/412

+++ internal/ceres/block_sparse_matrix.h
@@ -79,7 +79,7 @@ class BlockSparseMatrix : public SparseMatrix {
 
   int num_rows()         const final { return num_rows_;     }
   int num_cols()         const final { return num_cols_;     }
-  int num_nonzeros()     const final { return num_nonzeros_; }
+  int64_t num_nonzeros() const final { return num_nonzeros_; }
   const double* values() const final { return values_.get(); }
   double* mutable_values()     final { return values_.get(); }
 
@@ -127,8 +127,8 @@ class BlockSparseMatrix : public SparseMatrix {
  private:
   int num_rows_;
   int num_cols_;
-  int num_nonzeros_;
-  int max_num_nonzeros_;
+  int64_t num_nonzeros_;
+  int64_t max_num_nonzeros_;
   std::unique_ptr<double[]> values_;
   std::unique_ptr<CompressedRowBlockStructure> block_structure_;
 };
