+++ internal/ceres/compressed_row_sparse_matrix.h
@@ -90,7 +90,7 @@ class CompressedRowSparseMatrix : public SparseMatrix {
   // double the peak memory usage.
   //
   // The storage type is set to UNSYMMETRIC.
-  CompressedRowSparseMatrix(int num_rows, int num_cols, int max_num_nonzeros);
+  CompressedRowSparseMatrix(int num_rows, int num_cols, int64_t max_num_nonzeros);
 
   // Build a square sparse diagonal matrix with num_rows rows and
   // columns. The diagonal m(i,i) = diagonal(i);
@@ -109,7 +109,7 @@ class CompressedRowSparseMatrix : public SparseMatrix {
   void ToTextFile(FILE* file) const final;
   int num_rows() const final { return num_rows_; }
   int num_cols() const final { return num_cols_; }
-  int num_nonzeros() const final { return rows_[num_rows_]; }
+  int64_t num_nonzeros() const final { return rows_[num_rows_]; }
   const double* values() const final { return &values_[0]; }
   double* mutable_values() final { return &values_[0]; }
 
@@ -126,7 +126,7 @@ class CompressedRowSparseMatrix : public SparseMatrix {
   CompressedRowSparseMatrix* Transpose() const;
 
   // Destructive array resizing method.
-  void SetMaxNumNonZeros(int num_nonzeros);
+  void SetMaxNumNonZeros(int64_t num_nonzeros);
 
   // Non-destructive array resizing method.
   void set_num_rows(const int num_rows) { num_rows_ = num_rows; }
